+++ app/src/renderer/modules/data-manager/components/ImportButton.tsx
@@ -3,24 +3,10 @@ import React from 'react';
 import { Button } from '@mui/material';
 import FileUploadIcon from '@mui/icons-material/FileUpload';
 
-import { importCSV } from '@renderer/store/data-manager/actions';
-
 // eslint-disable-next-line import/named
 import { connect, ConnectedProps } from 'react-redux';
 import { RootState } from '@renderer/store';
-
-const mapState = (state: RootState) => ({
-  page: state.dataManager.page,
-  pageSize: state.dataManager.pageSize,
-});
-
-const mapDispatch = {
-  importCSV,
-};
-
-const connector = connect(mapState, mapDispatch);
-
-type PropsFromRedux = ConnectedProps<typeof connector>;
+import { importCSV } from '@renderer/store/data-manager/actions';
 
 const ImportButton: React.FC<PropsFromRedux> = props => {
   const importData = () => {
@@ -34,4 +20,18 @@ const ImportButton: React.FC<PropsFromRedux> = props => {
   );
 };
 
+// <redux>
+const mapState = (state: RootState) => ({
+  page: state.dataManager.page,
+  pageSize: state.dataManager.pageSize,
+});
+
+const mapDispatch = {
+  importCSV,
+};
+
+const connector = connect(mapState, mapDispatch);
+type PropsFromRedux = ConnectedProps<typeof connector>;
+
 export default connector(ImportButton);
+// </redux>
