+++ app/src/renderer/modules/data-manager/styles.ts
@@ -1,0 +1,16 @@
+export const noDataLoadedMessageStyles = {
+  height: '100%',
+  p: 2,
+  pt: '5%',
+  display: 'flex',
+  justifyContent: 'center',
+} as const;
+
+export const dataFrameWrapperStyles = {
+  flex: 1,
+  display: 'flex',
+  flexDirection: 'column',
+  boxSizing: 'border-box',
+  p: 1.5,
+  pt: 0,
+} as const;
